Remove section about working hours
So I've realised over the last few months that this was not a
personality quirk, but rather a bad habit and a step on the path to
burnout. It's true that programming helps me de-stress, but it doesn't
need to be work, and I was probably setting a bad example for my reports
and creating unreasonable expectations of the work we could get done
as ICs and as a team.